refactor: move PsiMetrics to sentinel-common for cross-crate reuse
PsiMetrics and parse_psi() are needed by both sentinel-sandbox (cgroups)
and sentinel-ebpf (monitoring). Move them to sentinel-common/src/psi.rs
and re-export from sentinel-sandbox to maintain backward compatibility.
